+++ CHANGELOG.md
@@ -5,12 +5,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [WIP]
+## [0.0.28] - 2022-12-12
+### Added
+- Added support for Pexels.
+    - Pexels requires a query and caption to function.
+- Added support for Unsplash.
+    - Unsplash does not require a query but does require a caption.
+    - Using Unsplash without a query will result in it using a random photo.
+- Added support for Emoji.
+    
+## [Released]
 ## [0.0.27] - 2022-11-25
 ### Added
 - Added ability to direct message users.
     - Direct messaging is only available through `follow(..., type=FollowType.Users)` and `like(..., type=LikeType.Users)`.
 
-## [Released]
 ## [0.0.26] - 2022-11-11
 ### Added
 - Added basic implementation to post media.